requirements.txt
# Core dependencies for Agentic RAG Application
docling>=2.58.0
ipywidgets>=8.1.7
langchain-community>=0.4.1
langchain-postgres>=0.0.16
langchain[openai]>=1.0.2
numpy>=2.3.4
pgvector>=0.3.6
psycopg>=3.2.12
psycopg2>=2.9.11
psycopg2-binary>=2.9.11
python-dotenv>=1.2.1
supabase>=2.22.4
vecs>=0.4.5

# Additional dependencies for document processing
transformers>=4.30.0

# Optional: Jupyter for running notebooks
jupyter>=1.0.0
notebook>=7.0.0
